Fix dart-side code
