+++ lib/src/dio/certificate_pinning_interceptor.dart
@@ -4,7 +4,7 @@ import 'package:dio/dio.dart';
 import 'package:flutter/services.dart';
 import 'package:http_certificate_pinning/http_certificate_pinning.dart';
 
-import 'exceptions.dart';
+import '../exceptions/exceptions.dart';
 
 class CertificatePinningInterceptor extends Interceptor {
   final List<String> _allowedSHAFingerprints;
