+++ lib/src/exceptions/exceptions.dart
